preparing access stage, refs #3625

+++ src/microsim/MSTransportableControl.cpp
@@ -239,10 +239,12 @@ MSTransportableControl::hasNonWaiting() const {
     return !myWaiting4Departure.empty() || myWaitingForVehicleNumber < myRunningNumber || myHaveNewWaiting;
 }
 
-int 
+
+int
 MSTransportableControl::getActiveCount() {
-    return myWaiting4Departure.size() + myRunningNumber - myWaitingForVehicleNumber;
+    return (int)myWaiting4Departure.size() + myRunningNumber - myWaitingForVehicleNumber;
 }
+
 
 void
 MSTransportableControl::abortWaitingForVehicle() {
